Hash field metadata to bypass field sync if not needed
This commit stores an MD5 hash of all field metadata received from the
database for a given table. Most of the time when sync runs, there
will have been no changes to the related fields of that table. By
hashing this and comparing that hash in subsequent sync runs, we can
avoid comparing stored field metadata with the field metadata that
comes back from the database driver. By avoiding this work, we
significantly increase sync performance and lower the impact of sync
on the rest of the system.

This hashing code only impacts the sync of fields as that was shown to
be the slowest part of the sync process. The other steps/stages of
sync/analyze are not affected by this change.